pyproject.toml
[tool.poetry]
name = "deus-use-case"
version = "0.1.0"
description = ""
authors = ["Joao Barroca"]
readme = "README.md"
packages = [
    {include = "src"},
]

[tool.poetry.dependencies]
python =  ">=3.9,<3.11,!=3.9.7"
pandas = "^2.1.1"
nltk = "^3.8.1"
pyldavis = "^3.4.1"
rank-eval = "^0.1.3"
chromadb = "^0.4.13"
sentence-transformers = "^2.2.2"
streamlit = "^1.27.1"
watchdog = "^3.0.0"
black = "^23.9.1"
ruff = "^0.0.291"
mypy = "^1.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
ignore_missing_imports = true
disable_error_code = ["type-abstract", "override"]

[tool.ruff]
line-length = 120
ignore-init-module-imports = true
[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
line-length = 120
